+++ aiotdlib/api/functions/get_external_link_info.py
@@ -12,7 +12,7 @@ from ..base_object import BaseObject
 
 class GetExternalLinkInfo(BaseObject):
     """
-    Returns information about an action to be done when the current user clicks an external link. Don't use this method for links from secret chats if link preview is disabled in secret chats
+    Returns information about an action to be done when the current user clicks an external link. Don't use this method for links from secret chats if web page preview is disabled in secret chats
     
     Params:
         link (:class:`str`)