+++ aiotdlib/api/functions/get_message_link.py
@@ -12,7 +12,7 @@ from ..base_object import BaseObject
 
 class GetMessageLink(BaseObject):
     """
-    Returns an HTTPS link to a message in a chat. Available only for already sent messages in supergroups and channels. This is an offline request
+    Returns an HTTPS link to a message in a chat. Available only for already sent messages in supergroups and channels, or if message.can_get_media_timestamp_links and a media timestamp link is generated. This is an offline request
     
     Params:
         chat_id (:class:`int`)
@@ -22,7 +22,7 @@ class GetMessageLink(BaseObject):
             Identifier of the message
         
         media_timestamp (:class:`int`)
-            If not 0, timestamp from which the video/audio/video note/voice note playing should start, in seconds. The media can be in the message content or in its link preview
+            If not 0, timestamp from which the video/audio/video note/voice note playing should start, in seconds. The media can be in the message content or in its web page preview
         
         for_album (:class:`bool`)
             Pass true to create a link for the whole media album