+++ aiotdlib/api/functions/get_message_public_forwards.py
@@ -12,7 +12,7 @@ from ..base_object import BaseObject
 
 class GetMessagePublicForwards(BaseObject):
     """
-    Returns forwarded copies of a channel message to different public channels. For optimal performance the number of returned messages is chosen by the library
+    Returns forwarded copies of a channel message to different public channels. For optimal performance, the number of returned messages is chosen by TDLib
     
     Params:
         chat_id (:class:`int`)
@@ -25,7 +25,7 @@ class GetMessagePublicForwards(BaseObject):
             Offset of the first entry to return as received from the previous request; use empty string to get first chunk of results
         
         limit (:class:`int`)
-            The maximum number of messages to be returned; must be positive and can't be greater than 100. Fewer messages may be returned than specified by the limit, even if the end of the list has not been reached
+            The maximum number of messages to be returned; must be positive and can't be greater than 100. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit
         
     """