+++ aiotdlib/api/functions/get_message_thread_history.py
@@ -12,7 +12,7 @@ from ..base_object import BaseObject
 
 class GetMessageThreadHistory(BaseObject):
     """
-    Returns messages in a message thread of a message. Can be used only if message.can_get_message_thread == true. Message thread of a channel message is in the channel's linked supergroup. The messages are returned in a reverse chronological order (i.e., in order of decreasing message_id). For optimal performance the number of returned messages is chosen by the library
+    Returns messages in a message thread of a message. Can be used only if message.can_get_message_thread == true. Message thread of a channel message is in the channel's linked supergroup. The messages are returned in a reverse chronological order (i.e., in order of decreasing message_id). For optimal performance, the number of returned messages is chosen by TDLib
     
     Params:
         chat_id (:class:`int`)
@@ -28,7 +28,7 @@ class GetMessageThreadHistory(BaseObject):
             Specify 0 to get results from exactly the from_message_id or a negative offset up to 99 to get additionally some newer messages
         
         limit (:class:`int`)
-            The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than or equal to -offset. Fewer messages may be returned than specified by the limit, even if the end of the message thread history has not been reached
+            The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than or equal to -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit
         
     """