+++ aiotdlib/api/functions/get_poll_voters.py
@@ -12,7 +12,7 @@ from ..base_object import BaseObject
 
 class GetPollVoters(BaseObject):
     """
-    Returns users voted for the specified option in a non-anonymous polls. For the optimal performance the number of returned users is chosen by the library
+    Returns users voted for the specified option in a non-anonymous polls. For optimal performance, the number of returned users is chosen by TDLib
     
     Params:
         chat_id (:class:`int`)
@@ -28,7 +28,7 @@ class GetPollVoters(BaseObject):
             Number of users to skip in the result; must be non-negative
         
         limit (:class:`int`)
-            The maximum number of users to be returned; must be positive and can't be greater than 50. Fewer users may be returned than specified by the limit, even if the end of the voter list has not been reached
+            The maximum number of users to be returned; must be positive and can't be greater than 50. For optimal performance, the number of returned users is chosen by TDLib and can be smaller than the specified limit, even if the end of the voter list has not been reached
         
     """