+++ aiotdlib/api/functions/get_trending_sticker_sets.py
@@ -12,14 +12,14 @@ from ..base_object import BaseObject
 
 class GetTrendingStickerSets(BaseObject):
     """
-    Returns a list of trending sticker sets. For the optimal performance the number of returned sticker sets is chosen by the library
+    Returns a list of trending sticker sets. For optimal performance, the number of returned sticker sets is chosen by TDLib
     
     Params:
         offset (:class:`int`)
             The offset from which to return the sticker sets; must be non-negative
         
         limit (:class:`int`)
-            The maximum number of sticker sets to be returned; must be non-negative. Fewer sticker sets may be returned than specified by the limit, even if the end of the list has not been reached
+            The maximum number of sticker sets to be returned; must be non-negative. For optimal performance, the number of returned sticker sets is chosen by TDLib and can be smaller than the specified limit, even if the end of the list has not been reached
         
     """