+++ aiotdlib/api/functions/search_chat_messages.py
@@ -14,7 +14,7 @@ from ..types import SearchMessagesFilter
 
 class SearchChatMessages(BaseObject):
     """
-    Searches for messages with given words in the chat. Returns the results in reverse chronological order, i.e. in order of decreasing message_id. Cannot be used in secret chats with a non-empty query (searchSecretMessages should be used instead), or without an enabled message database. For optimal performance the number of returned messages is chosen by the library
+    Searches for messages with given words in the chat. Returns the results in reverse chronological order, i.e. in order of decreasing message_id. Cannot be used in secret chats with a non-empty query (searchSecretMessages should be used instead), or without an enabled message database. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit
     
     Params:
         chat_id (:class:`int`)
@@ -33,7 +33,7 @@ class SearchChatMessages(BaseObject):
             Specify 0 to get results from exactly the from_message_id or a negative offset to get the specified message and some newer messages
         
         limit (:class:`int`)
-            The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than -offset. Fewer messages may be returned than specified by the limit, even if the end of the message history has not been reached
+            The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit
         
         filter_ (:class:`SearchMessagesFilter`)
             Filter for message content in the search results