+++ aiotdlib/api/functions/search_secret_messages.py
@@ -13,7 +13,7 @@ from ..types import SearchMessagesFilter
 
 class SearchSecretMessages(BaseObject):
     """
-    Searches for messages in secret chats. Returns the results in reverse chronological order. For optimal performance the number of returned messages is chosen by the library
+    Searches for messages in secret chats. Returns the results in reverse chronological order. For optimal performance, the number of returned messages is chosen by TDLib
     
     Params:
         chat_id (:class:`int`)
@@ -26,7 +26,7 @@ class SearchSecretMessages(BaseObject):
             Offset of the first entry to return as received from the previous request; use empty string to get first chunk of results
         
         limit (:class:`int`)
-            The maximum number of messages to be returned; up to 100. Fewer messages may be returned than specified by the limit, even if the end of the message history has not been reached
+            The maximum number of messages to be returned; up to 100. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit
         
         filter_ (:class:`SearchMessagesFilter`)
             A filter for message content in the search results