+++ aiotdlib/api/types/internal_link_type.py
@@ -155,7 +155,7 @@ class InternalLinkTypeFilterSettings(InternalLinkType):
 
 class InternalLinkTypeGame(InternalLinkType):
     """
-    The link is a link to a game. Call searchPublicChat with the given bot username, check that the user is a bot, ask the current user to select a group to send the game, and then call sendMessage with inputMessageGame
+    The link is a link to a game. Call searchPublicChat with the given bot username, check that the user is a bot, ask the current user to select a chat to send the game, and then call sendMessage with inputMessageGame
     
     Params:
         bot_username (:class:`str`)