+++ aiotdlib/api/types/message_link_info.py
@@ -28,7 +28,7 @@ class MessageLinkInfo(BaseObject):
             If found, the linked message; may be null
         
         media_timestamp (:class:`int`)
-            Timestamp from which the video/audio/video note/voice note playing should start, in seconds; 0 if not specified. The media can be in the message content or in its link preview
+            Timestamp from which the video/audio/video note/voice note playing should start, in seconds; 0 if not specified. The media can be in the message content or in its web page preview
         
         for_album (:class:`bool`)
             True, if the whole media album to which the message belongs is linked